more tweaks in users api routes but still needs attention

+++ server/users.js
@@ -35,9 +35,7 @@ module.exports = require('express').Router()
     selfOrAdmin,
     (req, res, next) => {
       User.findById(req.params.id, {include: [{model: Orders, include: [{model: OrderItem, include: [Products]}]}]})
-      .then(user => {
-        return res.json(user)
-      })
+      .then(user => res.json(user))
       .catch(next)
     })
   .get('/:id/orders/:orderId', mustBeLoggedIn, // this could be handled in orders route with conditional before sending response -- KHCL */
@@ -52,25 +50,19 @@ module.exports = require('express').Router()
       }).then((order) => res.json(order))
       .catch(next)
     })
-    // MPM I'm confused by this whole .put situation
-    .put('/:id', selfOnly, (req, res, next) => { // security -- KHCL
-      // MPM fix .update, but we'll also need to do a selfOrAdmin check, right?
-      const userId = req.params.id
-      const data = req.body // can they change their own isAdmin? -- KHCL
-      User.update(
-        {data},
-        {where: {id: userId}}
-      ).spread((affectedUsers) /* you don't use this varialbe so why define it? -- KHCL */ => User.findById(userId)).then((user) => { // chain .then in the same way visually you do in the rest of the file -- KHCL
-        res.json(user) // 201 created doesn't make sense here -- KHCL
-      }).catch(next)
-    })
+  .put('/:id', selfOnly, (req, res, next) => { // instead of selfOnly, use selfOrAdmin???
+    // MPM fix User.update, but we'll also need to do a selfOrAdmin check, right?
+    const userId = req.params.id
+    const data = req.body // can they change their own isAdmin? -- KHCL
+    User.update({data}, { where: {id: userId} })
+    .spread((affectedUsers) => User.findById(userId))
+    .then((user) => res.json(user))
+    .catch(next)
+  })
+  // fix this one too
   .delete('/:id', forbidden('Only an admin can do that'), (req, res, next) => { // security -- kHCL
     const userId = req.params.id
-    User.destroy({ // look into cascade option to replace all other destroys (those should be happening in parallel anyway) -- KHCL
-      where: {
-        id: userId
-      }
-    })
+    User.destroy({ where: {id: userId} })
       .then((affectedRows) => {
         if (affectedRows === 0) {
           res.status(404)
